+++ src/OpenTelemetry.Instrumentation.Http/Implementation/HttpWebRequestActivitySource.netfx.cs
@@ -23,6 +23,7 @@ using System.Reflection;
 using System.Reflection.Emit;
 using System.Runtime.CompilerServices;
 using OpenTelemetry.Context.Propagation;
+using OpenTelemetry.Internal;
 using OpenTelemetry.Trace;
 using static OpenTelemetry.Internal.HttpSemanticConventionHelper;
 
@@ -153,7 +154,18 @@ internal static class HttpWebRequestActivitySource
             // see the spec https://github.com/open-telemetry/semantic-conventions/blob/v1.21.0/docs/http/http-spans.md
             if (tracingEmitNewAttributes)
             {
-                activity.SetTag(SemanticConventions.AttributeHttpRequestMethod, request.Method);
+                if (RequestMethodHelper.KnownMethods.TryGetValue(request.Method, out var httpMethod))
+                {
+                    activity.SetTag(SemanticConventions.AttributeHttpRequestMethod, httpMethod);
+                }
+                else
+                {
+                    // Set to default "_OTHER" as per spec.
+                    // https://github.com/open-telemetry/semantic-conventions/blob/v1.22.0/docs/http/http-spans.md#common-attributes
+                    activity.SetTag(SemanticConventions.AttributeHttpRequestMethod, "_OTHER");
+                    activity.SetTag(SemanticConventions.AttributeHttpRequestMethodOriginal, request.Method);
+                }
+
                 activity.SetTag(SemanticConventions.AttributeServerAddress, request.RequestUri.Host);
                 if (!request.RequestUri.IsDefaultPort)
                 {
@@ -206,6 +218,31 @@ internal static class HttpWebRequestActivitySource
     }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    private static string GetErrorType(Exception exception)
+    {
+        if (exception is WebException wexc)
+        {
+            // TODO: consider other Status values from
+            // https://learn.microsoft.com/dotnet/api/system.net.webexceptionstatus?view=netframework-4.6.2
+            return wexc.Status switch
+            {
+                WebExceptionStatus.NameResolutionFailure => "name_resolution_failure",
+                WebExceptionStatus.ConnectFailure => "connect_failure",
+                WebExceptionStatus.SendFailure => "send_failure",
+                WebExceptionStatus.RequestCanceled => "request_cancelled",
+                WebExceptionStatus.TrustFailure => "trust_failure",
+                WebExceptionStatus.SecureChannelFailure => "secure_channel_failure",
+                WebExceptionStatus.ServerProtocolViolation => "server_protocol_violation",
+                WebExceptionStatus.Timeout => "timeout",
+                WebExceptionStatus.MessageLengthLimitExceeded => "message_length_limit_exceeded",
+                _ => wexc.GetType().FullName,
+            };
+        }
+
+        return exception.GetType().FullName;
+    }
+
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
     private static void AddExceptionTags(Exception exception, Activity activity, out HttpStatusCode? statusCode)
     {
         Debug.Assert(activity != null, "Activity must not be null");
@@ -218,57 +255,30 @@ internal static class HttpWebRequestActivitySource
         }
 
         ActivityStatusCode status;
-        string exceptionMessage = null;
 
-        if (exception is WebException wexc)
+        if (exception is WebException wexc && wexc.Response is HttpWebResponse response)
         {
-            if (wexc.Response is HttpWebResponse response)
+            statusCode = response.StatusCode;
+
+            if (tracingEmitOldAttributes)
             {
-                statusCode = response.StatusCode;
-
-                if (tracingEmitOldAttributes)
-                {
-                    activity.SetTag(SemanticConventions.AttributeHttpStatusCode, (int)statusCode);
-                }
-
-                if (tracingEmitNewAttributes)
-                {
-                    activity.SetTag(SemanticConventions.AttributeHttpResponseStatusCode, (int)statusCode);
-                }
-
-                status = SpanHelper.ResolveSpanStatusForHttpStatusCode(activity.Kind, (int)statusCode);
+                activity.SetTag(SemanticConventions.AttributeHttpStatusCode, (int)statusCode);
             }
-            else
+
+            if (tracingEmitNewAttributes)
             {
-                switch (wexc.Status)
-                {
-                    case WebExceptionStatus.Timeout:
-                    case WebExceptionStatus.RequestCanceled:
-                        status = ActivityStatusCode.Error;
-                        break;
-                    case WebExceptionStatus.SendFailure:
-                    case WebExceptionStatus.ConnectFailure:
-                    case WebExceptionStatus.SecureChannelFailure:
-                    case WebExceptionStatus.TrustFailure:
-                    case WebExceptionStatus.ServerProtocolViolation:
-                    case WebExceptionStatus.MessageLengthLimitExceeded:
-                        status = ActivityStatusCode.Error;
-                        exceptionMessage = exception.Message;
-                        break;
-                    default:
-                        status = ActivityStatusCode.Error;
-                        exceptionMessage = exception.Message;
-                        break;
-                }
+                activity.SetTag(SemanticConventions.AttributeHttpResponseStatusCode, (int)statusCode);
             }
+
+            status = SpanHelper.ResolveSpanStatusForHttpStatusCode(activity.Kind, (int)statusCode);
         }
         else
         {
             status = ActivityStatusCode.Error;
-            exceptionMessage = exception.Message;
         }
 
-        activity.SetStatus(status, exceptionMessage);
+        activity.SetStatus(status);
+
         if (TracingOptions.RecordException)
         {
             activity.RecordException(exception);
@@ -386,6 +396,7 @@ internal static class HttpWebRequestActivitySource
     private static void ProcessResult(IAsyncResult asyncResult, AsyncCallback asyncCallback, Activity activity, object result, bool forceResponseCopy, HttpWebRequest request, long startTimestamp)
     {
         HttpStatusCode? httpStatusCode = null;
+        string errorType = null;
 
         // Activity may be null if we are not tracing in these cases:
         // 1. No listeners
@@ -401,6 +412,7 @@ internal static class HttpWebRequestActivitySource
         {
             if (result is Exception ex)
             {
+                errorType = GetErrorType(ex);
                 if (activity != null)
                 {
                     AddExceptionTags(ex, activity, out httpStatusCode);
@@ -445,11 +457,22 @@ internal static class HttpWebRequestActivitySource
 
                     httpStatusCode = response.StatusCode;
                 }
+
+                if (SpanHelper.ResolveSpanStatusForHttpStatusCode(ActivityKind.Client, (int)httpStatusCode.Value) == ActivityStatusCode.Error)
+                {
+                    // override the errorType to statusCode for failures.
+                    errorType = TelemetryHelper.GetStatusCodeString(httpStatusCode.Value);
+                }
             }
         }
         catch (Exception ex)
         {
             HttpInstrumentationEventSource.Log.FailedProcessResult(ex);
+        }
+
+        if (tracingEmitNewAttributes && errorType != null)
+        {
+            activity?.SetTag(SemanticConventions.AttributeErrorType, errorType);
         }
 
         activity?.Stop();
@@ -495,7 +518,17 @@ internal static class HttpWebRequestActivitySource
             {
                 TagList tags = default;
 
-                tags.Add(SemanticConventions.AttributeHttpRequestMethod, request.Method);
+                if (RequestMethodHelper.KnownMethods.TryGetValue(request.Method, out var httpMethod))
+                {
+                    tags.Add(new KeyValuePair<string, object>(SemanticConventions.AttributeHttpRequestMethod, httpMethod));
+                }
+                else
+                {
+                    // Set to default "_OTHER" as per spec.
+                    // https://github.com/open-telemetry/semantic-conventions/blob/v1.22.0/docs/http/http-spans.md#common-attributes
+                    tags.Add(new KeyValuePair<string, object>(SemanticConventions.AttributeHttpRequestMethod, "_OTHER"));
+                }
+
                 tags.Add(SemanticConventions.AttributeServerAddress, request.RequestUri.Host);
                 tags.Add(SemanticConventions.AttributeUrlScheme, request.RequestUri.Scheme);
                 tags.Add(SemanticConventions.AttributeNetworkProtocolVersion, HttpTagHelper.GetFlavorTagValueFromProtocolVersion(request.ProtocolVersion));
@@ -507,6 +540,11 @@ internal static class HttpWebRequestActivitySource
                 if (httpStatusCode.HasValue)
                 {
                     tags.Add(SemanticConventions.AttributeHttpResponseStatusCode, (int)httpStatusCode.Value);
+                }
+
+                if (errorType != null)
+                {
+                    tags.Add(SemanticConventions.AttributeErrorType, errorType);
                 }
 
                 HttpClientRequestDuration.Record(durationS, tags);