+++ src/OpenTelemetry/Metrics/AggregatorStore.cs
@@ -372,10 +372,19 @@ internal sealed class AggregatorStore
 
     private static double[] FindDefaultHistogramBounds(in MetricStreamIdentity metricStreamIdentity)
     {
-        if (metricStreamIdentity.Unit == "s" && Metric.DefaultHistogramBoundMappings
-            .Contains((metricStreamIdentity.MeterName, metricStreamIdentity.InstrumentName)))
+        if (metricStreamIdentity.Unit == "s")
         {
-            return Metric.DefaultHistogramBoundsSeconds;
+            if (Metric.DefaultHistogramBoundShortMappings
+                .Contains((metricStreamIdentity.MeterName, metricStreamIdentity.InstrumentName)))
+            {
+                return Metric.DefaultHistogramBoundsShortSeconds;
+            }
+
+            if (Metric.DefaultHistogramBoundLongMappings
+                .Contains((metricStreamIdentity.MeterName, metricStreamIdentity.InstrumentName)))
+            {
+                return Metric.DefaultHistogramBoundsLongSeconds;
+            }
         }
 
         return Metric.DefaultHistogramBounds;