+++ src/OpenTelemetry/Resources/ResourceBuilder.cs
@@ -59,7 +59,7 @@ public class ResourceBuilder
     /// <summary>
     /// Creates a <see cref="ResourceBuilder"/> instance with default attributes
     /// added. See <a
-    /// href="https://github.com/open-telemetry/opentelemetry-specification/tree/main/specification/resource/semantic_conventions#semantic-attributes-with-sdk-provided-default-value">resource
+    /// href="https://github.com/open-telemetry/semantic-conventions/blob/main/docs/resource/README.md#semantic-attributes-with-sdk-provided-default-value">resource
     /// semantic conventions</a> for details.
     /// Additionally it adds resource attributes parsed from OTEL_RESOURCE_ATTRIBUTES, OTEL_SERVICE_NAME environment variables
     /// to a <see cref="ResourceBuilder"/> following the <a